+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Add `RedisMigration` type which allows to write migrations in Go code
 - Add `RedisMigrator` type which implements core migration logic
+- Add support for cancelling migrations via the given context